Reorganize tests and add library folder
