+++ test/library/MerkleProof.test.js
@@ -1,8 +1,8 @@
 
-import MerkleTree from './helpers/merkleTree.js';
+import MerkleTree from '../helpers/merkleTree.js';
 import { sha3, bufferToHex } from 'ethereumjs-util';
 
-var MerkleProof = artifacts.require('./MerkleProof.sol');
+var MerkleProof = artifacts.require('MerkleProof.sol');
 
 contract('MerkleProof', function (accounts) {
   let merkleProof;
